API: wire auth + permissions into document endpoints
Extract document CRUD and export/import into separate route modules.
Wire auth middleware globally on /api routes and permission checks
on each document endpoint. Auto-grant owner role on document creation.
