+++ modules/api/internal/export-routes.ts
@@ -1,0 +1,63 @@
+/** Contract: contracts/api/rules.md */
+
+import { Router, type Request, type Response } from 'express';
+import { getDocument } from '../../storage/internal/pg.ts';
+import { getDocumentForExport } from '../../convert/internal/converter.ts';
+import type { PermissionsModule } from '../../permissions/index.ts';
+
+export type ExportRoutesOptions = {
+  permissions: PermissionsModule;
+};
+
+/**
+ * Mount export/import routes onto a router.
+ * Each route enforces authentication and permission checks.
+ */
+export function createExportRoutes(opts: ExportRoutesOptions): Router {
+  const router = Router();
+  const { permissions } = opts;
+
+  // Export document — requires read permission
+  router.post('/:id/export', permissions.require('read'), async (req: Request, res: Response) => {
+    const format = req.body?.format;
+    if (!format || !['html', 'text'].includes(format)) {
+      res.status(400).json({ error: 'format must be "html" or "text"' });
+      return;
+    }
+    const meta = await getDocumentForExport(req.params.id);
+    if (!meta) {
+      res.status(404).json({ error: 'Document not found' });
+      return;
+    }
+    const content = req.body?.content;
+    if (!content && content !== '') {
+      res.status(400).json({ error: 'content is required (client-side export)' });
+      return;
+    }
+    const ext = format === 'html' ? 'html' : 'txt';
+    const filename = `${meta.title || 'document'}.${ext}`;
+    const contentType = format === 'html'
+      ? 'text/html; charset=utf-8'
+      : 'text/plain; charset=utf-8';
+    res.setHeader('Content-Type', contentType);
+    res.setHeader('Content-Disposition', `attachment; filename="${encodeURIComponent(filename)}"`);
+    res.send(content);
+  });
+
+  // Import HTML into document — requires write permission
+  router.post('/:id/import', permissions.require('write'), async (req: Request, res: Response) => {
+    const html = req.body?.html;
+    if (!html) {
+      res.status(400).json({ error: 'html content is required' });
+      return;
+    }
+    const doc = await getDocument(req.params.id);
+    if (!doc) {
+      res.status(404).json({ error: 'Document not found' });
+      return;
+    }
+    res.json({ ok: true, html, documentId: req.params.id });
+  });
+
+  return router;
+}
