+++ modules/api/internal/server.ts
@@ -1,27 +1,36 @@
 /** Contract: contracts/api/rules.md */
 import { createServer } from 'node:http';
-import { randomUUID } from 'node:crypto';
 import express from 'express';
 import { resolve, dirname } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { createCollabServer } from '../../collab/internal/server.ts';
-import {
-  listDocuments,
-  createDocument,
-  getDocument,
-  deleteDocument,
-  updateDocumentTitle,
-} from '../../storage/internal/pg.ts';
-import { getDocumentForExport } from '../../convert/internal/converter.ts';
 import { getRedisClient, disconnectRedis } from './redis.ts';
 import { idempotencyMiddleware } from './idempotency.ts';
 import { createConvertRoutes } from './convert-routes.ts';
+import { createAuth } from '../../auth/index.ts';
+import { createPermissions } from '../../permissions/index.ts';
+import { createDocumentRoutes } from './document-routes.ts';
+import { createExportRoutes } from './export-routes.ts';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
 export function startServer(port = 3000) {
   const app = express();
   const { handleUpgrade } = createCollabServer();
+
+  // Wire auth module (dev mode uses bypass verifiers)
+  const auth = createAuth({
+    serviceAccountStore: { findByHashedKey: async () => null },
+    serviceAccountStorage: {
+      save: async () => {},
+      findById: async () => null,
+      delete: async () => {},
+    },
+    publicPaths: ['/api/health'],
+  });
+
+  // Wire permissions module
+  const permissions = createPermissions();
 
   app.use(express.json());
 
@@ -36,96 +45,19 @@ export function startServer(port = 3000) {
   const publicDir = resolve(__dirname, '../../app/internal/public');
   app.use(express.static(publicDir));
 
-  // Health check
+  // Auth middleware on all /api routes (except public paths)
+  app.use('/api', auth.middleware);
+
+  // Health check (public, skipped by auth middleware)
   app.get('/api/health', (_req, res) => {
     res.json({ status: 'ok', version: '0.1.0' });
   });
 
-  // List documents
-  app.get('/api/documents', async (_req, res) => {
-    const docs = await listDocuments();
-    res.json(docs);
-  });
+  // Document CRUD with permission checks
+  app.use('/api/documents', createDocumentRoutes({ permissions }));
 
-  // Create document
-  app.post('/api/documents', async (req, res) => {
-    const title = req.body?.title || 'Untitled';
-    const id = randomUUID();
-    const doc = await createDocument(id, title);
-    res.status(201).json(doc);
-  });
-
-  // Get document
-  app.get('/api/documents/:id', async (req, res) => {
-    const doc = await getDocument(req.params.id);
-    if (!doc) {
-      res.status(404).json({ error: 'Document not found' });
-      return;
-    }
-    res.json(doc);
-  });
-
-  // Update document title
-  app.patch('/api/documents/:id', async (req, res) => {
-    const { title } = req.body;
-    if (!title) {
-      res.status(400).json({ error: 'title is required' });
-      return;
-    }
-    await updateDocumentTitle(req.params.id, title);
-    res.json({ ok: true });
-  });
-
-  // Delete document
-  app.delete('/api/documents/:id', async (req, res) => {
-    const deleted = await deleteDocument(req.params.id);
-    if (!deleted) {
-      res.status(404).json({ error: 'Document not found' });
-      return;
-    }
-    res.json({ ok: true });
-  });
-
-  // Export document (MVP: returns metadata; actual content comes client-side)
-  app.post('/api/documents/:id/export', async (req, res) => {
-    const format = req.body?.format;
-    if (!format || !['html', 'text'].includes(format)) {
-      res.status(400).json({ error: 'format must be "html" or "text"' });
-      return;
-    }
-    const meta = await getDocumentForExport(req.params.id);
-    if (!meta) {
-      res.status(404).json({ error: 'Document not found' });
-      return;
-    }
-    // For MVP, the client sends content along with the request
-    const content = req.body?.content;
-    if (!content && content !== '') {
-      res.status(400).json({ error: 'content is required (client-side export)' });
-      return;
-    }
-    const filename = `${meta.title || 'document'}.${format === 'html' ? 'html' : 'txt'}`;
-    const contentType = format === 'html' ? 'text/html; charset=utf-8' : 'text/plain; charset=utf-8';
-    res.setHeader('Content-Type', contentType);
-    res.setHeader('Content-Disposition', `attachment; filename="${encodeURIComponent(filename)}"`);
-    res.send(content);
-  });
-
-  // Import HTML file into document (MVP: accept raw HTML in body)
-  app.post('/api/documents/:id/import', async (req, res) => {
-    const html = req.body?.html;
-    if (!html) {
-      res.status(400).json({ error: 'html content is required' });
-      return;
-    }
-    const doc = await getDocument(req.params.id);
-    if (!doc) {
-      res.status(404).json({ error: 'Document not found' });
-      return;
-    }
-    // For MVP, we return the HTML and let the client set it via editor.commands.setContent()
-    res.json({ ok: true, html, documentId: req.params.id });
-  });
+  // Export/import routes with permission checks
+  app.use('/api/documents', createExportRoutes({ permissions }));
 
   const httpServer = createServer(app);
